sauto3/m2/requirements.txt
websocket-client
requests
six
py-radius
